requirements.txt
annotated-types
anyio
certifi==2024.12.14
click==8.1.8
colorama
distro
dnspython==2.6.1
email_validator==2.2.0
exceptiongroup
fastapi
fastapi-cli==0.0.7
greenlet==3.1.1
h11
httpcore
httptools==0.6.4
httpx
idna
Jinja2==3.1.5
jiter
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
openai
pydantic
pydantic_core
Pygments==2.18.0
python-dotenv==1.0.1
python-multipart==0.0.20
PyYAML==6.0.2
rich==13.9.4
rich-toolkit==0.12.0
shellingham==1.5.4
sniffio
SQLAlchemy==2.0.36
starlette
tqdm
typer==0.15.1
typing_extensions
uvicorn==0.33.0
watchfiles==0.24.0
websockets==13.1
